3305:   boiler plate code for feTurbulence:   *reading attributes and setting default values
  same for some more attributes on convolve matrix filter

  alphabetical ordering on Makefile_insert and on nr-filter.cpp #includes

+++ src/sp-feconvolvematrix.h
@@ -30,7 +30,7 @@ struct SPFeConvolveMatrix : public SPFilterPrimitive {
     std::vector<gdouble> kernelMatrix;
     double divisor, bias;
     int targetX, targetY;
-    unsigned short edgeMode;
+    int edgeMode;
     NumberOptNumber kernelUnitLength;
     bool preserveAlpha;
 };
